// front-end-angular/src/app/models/company.model.ts
export interface Company {
    id: number;
    ra_social: string;
    ruc: string;
    address: string;
    status: string;
    phone: string;
    user_id: number;
    email: string;
    image: string | null
    image_url: string | null;
    created_at: string | null;
    updated_at: string | null;
}